finish changes

+++ controller/task.controller.js
@@ -3,13 +3,16 @@ const db = require('../db/connect')
 class TaskController {
     async createTask(req, res) {
         const {task} = req.body;
+        if (task.length > 255){
+            return res.status(400).send('Bad Request')
+        }
         try {
             const newTask = await db.query('INSERT INTO todo (task) values ($1) RETURNING *', [task])
-            res.json(newTask.rows[0])
+            return res.json(newTask.rows[0])
         }
         catch (e) {
             console.log(e)
-            res.status(500)
+            return res.status(500)
         }
     }
 
@@ -18,23 +21,27 @@ class TaskController {
         try {
             const tasks = await db.query('SELECT * FROM todo')
             console.log(tasks)
-            res.json(tasks.rows)
+            return res.json(tasks.rows)
         }
         catch (e) {
             console.log(e)
-            res.status(500)
+            return res.status(500)
         }
     }
 
     async deleteTask(req, res) {
         const id = req.params.id
+
+        if (Math.sign(id) === -1 || isNaN(Math.sign(id))){
+            return res.status(400).send('Bad Request')
+        }
         try {
             const task = await db.query('DELETE FROM todo where id = $1', [id])
-            res.json(task.rows[0])
+            return res.json(task.rows[0])
         }
         catch (e) {
             console.log(e)
-            res.status(500)
+            return res.status(500)
         }
     }
 }
